requirements.txt
streamlit
psycopg2-binary
python-dotenv
fastapi
uvicorn